Change default behavior of 'javaHome' option
This option previously meant 'Use $JAVA_HOME' if this option is not set
at all. It now means 'use $JAVA_HOME' if this option is not set OR if
the environment variable it's set to is not defined at runtime.

This is useful in environments where a default JDK is preferred, but
init should not fail should this JDK be absent.